Refactor project index helpers into dedicated modules
